fix: split cryptography module in different repositories
