+++ src/domain/delivery/application/cryptography/encrypter.ts
@@ -1,8 +1,5 @@
 abstract class Encrypter {
   abstract encrypt(payload: Record<string, unknown>): Promise<string>;
-
-  abstract hash(plainText: string): Promise<string>;
-  abstract compare(plainText: string, hash: string): Promise<boolean>;
 }
 
 export { Encrypter };
