+++ src/domain/delivery/application/cryptography/hash-comparer.ts
@@ -1,0 +1,5 @@
+abstract class HashComparer {
+  abstract compare(plainText: string, hash: string): Promise<boolean>;
+}
+
+export { HashComparer };
