+++ src/domain/delivery/application/cryptography/hash-generator.ts
@@ -1,0 +1,5 @@
+abstract class HashGenerator {
+  abstract hash(plainText: string): Promise<string>;
+}
+
+export { HashGenerator };
