+++ src/infra/cryptography/jwt-encrypter.ts
@@ -1,8 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 
+import { Encrypter } from '@/domain/delivery/application/cryptography/encrypter';
+
 @Injectable()
-class JwtEncryter {
+class JwtEncryter implements Encrypter {
   private jwtService: JwtService;
   constructor(jwtService: JwtService) {
     this.jwtService = jwtService;
