+++ tests/cryptography/fake-encrypter.ts
@@ -4,15 +4,6 @@ class FakeEncrypter implements Encrypter {
   async encrypt(payload: Record<string, unknown>) {
     return JSON.stringify(payload);
   }
-
-  async hash(plainText: string) {
-    return plainText.concat('-hashed');
-  }
-
-  async compare(plainText: string, hash: string) {
-    const hashedPlainText = await this.hash(plainText);
-    return hashedPlainText === hash;
-  }
 }
 
 export { FakeEncrypter };
